[#42] Automtic PID file path definition
In the case the user does not set the `pidfile` configuration
parameter, the software should define one path itself.
In particular the change makes the `pidfile` equal to the concatenation
of the `unix_socket_dir` (mandatory) and the string `pgmoneta.<host>.pid`
where `<host>` is the value of the `host` configuration parameter
or `all` if `host = *`.

Changed the `pidfile` variable size.

Updated the documentation and the AUTHORS file.

Close #42.

+++ src/main.c
@@ -1712,8 +1712,28 @@ create_pidfile(void)
 
    config = (struct configuration*)shmem;
 
+   if (strlen(config->pidfile) == 0)
+   {
+      // no pidfile set, use a default one
+      snprintf(config->pidfile, sizeof(config->pidfile), "%s/pgmoneta.%s.pid",
+               config->unix_socket_dir,
+               !strncmp(config->host,"*", sizeof(config->host)) ? "all" : config->host );
+      pgmoneta_log_debug("PID file automatically set to: [%s]", config->pidfile);
+   }
+   
    if (strlen(config->pidfile) > 0)
    {
+
+       if (strlen(config->pidfile) > 0)
+       {
+          // check pidfile is not there
+          if (access(config->pidfile, F_OK) == 0)
+          {
+             pgmoneta_log_fatal("PID file [%s] exists, is there another instance running ?", config->pidfile);
+             goto error;
+          }
+       }
+       
       pid = getpid();
 
       fd = open(config->pidfile, O_WRONLY | O_CREAT | O_EXCL, 0644);
@@ -1749,7 +1769,7 @@ remove_pidfile(void)
 
    config = (struct configuration*)shmem;
 
-   if (strlen(config->pidfile) > 0)
+   if (strlen(config->pidfile) > 0 && access(config->pidfile, F_OK) == 0)
    {
       unlink(config->pidfile);
    }
